[diffusion] Relocate conditioning expansion utility
