+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/denoising.py
@@ -754,9 +754,6 @@ class DenoisingStage(PipelineStage, RolloutDenoisingMixin):
         Returns:
             A context object containing the invariant state for the denoising loop.
         """
-        # This is the last shared boundary before model-specific conditioning
-        # kwargs are derived. The base config is a no-op; audited model families
-        # explicitly opt in to any prompt-to-sample batch normalization.
         server_args.pipeline_config.expand_conditioning_to_sample_batch(batch)
 
         assert self.transformer is not None
