+++ python/sglang/multimodal_gen/runtime/utils/condition_expansion.py
